Fix support for multiple registers
This commit fixes support for working with multiple registers in
mapomatic. To do this instead of working with bit indices (which is the
more natural way to think about them) this instead works with Qiskit bit
objects which are standalone and don't have an order outside of the
circuit. Whenever an order is needed the QuantumCircuit.find_bit() api is
used to get the bit's index and uses that. This then adds tests with
multiple registers to verify this works.