Deduplicate requirements lists
This commit deduplicates the requirements list so that moving forward we
only need to update the requirements.txt list for any changes in
requirements. Previously we had duplicate lists in the requirements.txt
file and the setup.py, however this was unecessary as the setup.py can
just read the contents of the requirements file and use that for the
requirements list.