Add tox configuration
To simplify running tests for local developers this commit adds a
a tox.ini file. Tox enables a single entrypoint that combines venv
management with running test commands. Local developers can leverage
this to automate the execution of tests and installing mapomatic in an
isolated venv when making local changes.