Marginally optimize interaction graph creation
This commit tweaks the interaction graph creation so that it marginally
more efficient. Instead of always having two calls to first create the
nodes and then add the edges this commit changes the interaction graph
creation to let retworkx create nodes at the same time it's adding
edges. This is ever so slightly faster because it's all done on the rust
side and we only need to copy and convert one object going over the
Python<->Rust FFI boundary.